Refactor core helpers to classes
