+++ armoreditor.php
@@ -2,6 +2,7 @@
 use Lotgd\SuAccess;
 use Lotgd\Nav\SuperuserNav;
 // translator ready
+use Lotgd\Forms;
 // addnews ready
 // mail ready
 /**
@@ -11,7 +12,6 @@ use Lotgd\Nav\SuperuserNav;
 * 
 */
 require_once("common.php");
-require_once("lib/showform.php");
 require_once("lib/http.php");
 
 SuAccess::check(SU_EDIT_EQUIPMENT);
@@ -47,7 +47,7 @@ if($op=="edit" || $op=="add"){
 	}
 	rawoutput("<form action='armoreditor.php?op=save&level=$armorlevel' method='POST'>");
 	addnav("","armoreditor.php?op=save&level=$armorlevel");
-	showform($armorarray,$row);
+	Forms::showForm($armorarray,$row);
 	rawoutput("</form>");
 }else if($op=="del"){
 	$sql = "DELETE FROM " . db_prefix("armor") . " WHERE armorid='$id'";
