+++ pages/about/about_setup.php
@@ -1,4 +1,5 @@
 <?php
+use Lotgd\Forms;
 addnav("About LoGD");
 addnav("About LoGD","about.php");
 addnav("Module Info","about.php?op=listmodules");
@@ -71,5 +72,5 @@ $form = array_merge($setup, $args['settings']);
 $form = array_merge($form, $useful);
 $vals = array_merge($localsettings, $args['values']);
 $vals = array_merge($vals, $useful_vals);
-showform($form,$vals,true);
+Forms::showForm($form,$vals,true);
 ?>
